requirements.txt
aiogram==3.12.0
httpx==0.27.2
python-dateutil==2.9.0.post0
deep-translator==1.11.4
